+++ tests/server/managers/AppSlashCommandManager.spec.ts
@@ -20,7 +20,7 @@ import type { ProxiedApp } from '../../../src/server/ProxiedApp';
 import { Room } from '../../../src/server/rooms/Room';
 import type { AppsEngineRuntime } from '../../../src/server/runtime/AppsEngineRuntime';
 import type { AppLogStorage } from '../../../src/server/storage';
-import type { DenoRuntimeSubprocessController } from '../../../src/server/runtime/AppsEngineDenoRuntime';
+import type { DenoRuntimeSubprocessController } from '../../../src/server/runtime/deno/AppsEngineDenoRuntime';
 
 export class AppSlashCommandManagerTestFixture {
     public static doThrow = false;
